cartcontroller init and fix cartController and cartService code

+++ src/main/java/shoppingMall/gupang/service/cart/CartServiceImpl.java
@@ -1,15 +1,12 @@
 package shoppingMall.gupang.service.cart;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import shoppingMall.gupang.web.controller.cart.dto.CartItemDto;
-import shoppingMall.gupang.web.controller.cart.dto.CartItemsMemberDto;
 import shoppingMall.gupang.domain.CartItem;
 import shoppingMall.gupang.domain.Item;
 import shoppingMall.gupang.domain.Member;
+import shoppingMall.gupang.exception.cart.CartItemNotMatchWithMemberException;
 import shoppingMall.gupang.exception.cart.LackOfCountException;
 import shoppingMall.gupang.exception.cart.NoCartItemException;
 import shoppingMall.gupang.exception.item.NoItemException;
@@ -17,6 +14,7 @@ import shoppingMall.gupang.exception.member.NoMemberException;
 import shoppingMall.gupang.repository.cartItem.CartRepository;
 import shoppingMall.gupang.repository.item.ItemRepository;
 import shoppingMall.gupang.repository.member.MemberRepository;
+import shoppingMall.gupang.web.controller.cart.dto.CartItemIdsDto;
 
 import java.util.List;
 import java.util.Optional;
@@ -32,19 +30,19 @@ public class CartServiceImpl implements CartService{
 
     @Transactional(readOnly = true)
     @Override
-    public List<CartItem> getAllCartItems(Long memberId) {
+    public List<CartItem> getAllCartItems(String memberEmail) {
 
-        Optional<Member> optionalMember = memberRepository.findById(memberId);
+        Optional<Member> optionalMember = memberRepository.findByEmail(memberEmail);
         Member member = optionalMember.orElse(null);
         if (member == null) {
             throw new NoMemberException("해당 회원이 없습니다.");
         }
 
-        return cartRepository.findCartItemsByMemberId(memberId);
+        return cartRepository.findCartItemsByMemberId(member.getId());
     }
 
     @Override
-    public void updateItemCount(Long cartItemId, int count) {
+    public List<CartItem> updateItemCount(String memberEmail, Long cartItemId, int count) {
 
         if (count < 1) {
             throw new LackOfCountException("최소 1이상 입력하세요");
@@ -56,12 +54,23 @@ public class CartServiceImpl implements CartService{
             throw new NoCartItemException("맞는 카트 상품이 없습니다.");
         }
 
+        Optional<Member> optionalMember = memberRepository.findByEmail(cartItem.getMember().getEmail());
+        Member member = optionalMember.orElse(null);
+        if (member == null) {
+            throw new NoMemberException("회원이 존재하지 않습니다.");
+        }
+
+        if (!memberEmail.equals(cartItem.getMember().getEmail())) {
+            throw new CartItemNotMatchWithMemberException("카트 상품의 소유자와 맞지 않습니다.");
+        }
+
         cartItem.updateItemCount(count);
+        return cartRepository.findCartItemsByMemberId(member.getId());
     }
 
     @Override
-    public void addCartItem(Long memberId, Long itemId, int count) {
-        Optional<Member> optionalMember = memberRepository.findById(memberId);
+    public void addCartItem(String memberEmail, Long itemId, int count) {
+        Optional<Member> optionalMember = memberRepository.findByEmail(memberEmail);
         Member member = optionalMember.orElse(null);
         if (member == null) {
             throw new NoMemberException("해당하는 멤버가 없습니다.");
@@ -78,35 +87,34 @@ public class CartServiceImpl implements CartService{
     }
 
     @Override
-    public List<CartItem> removeCartItems(CartItemsMemberDto cartItemsMemberDto) {
+    public List<CartItem> removeCartItems(String memberEmail, CartItemIdsDto cartItemIdsDto) {
 
-        for (CartItemDto dto : cartItemsMemberDto.getCartItemIds()) {
-            Optional<CartItem> optionalCartItem = cartRepository.findById(dto.getCartItemId());
+        /*
+            - cartItems 돌면서 memberEmail과 cartItem의 memberEmail과 맞지 않는 케이스가 있다면
+            - throw exception한다.
+         */
+        for (Long cartItemId : cartItemIdsDto.getCartItemIds()) {
+            Optional<CartItem> optionalCartItem = cartRepository.findById(cartItemId);
             CartItem cartItem = optionalCartItem.orElse(null);
             if (cartItem == null) {
                 throw new NoCartItemException("해당하는 카트 상품이 없습니다.");
             }
+
+            // 세션의 이메일과 cartItem의 멤버의 이메일 비교
+            String cartItemMemberEmail = cartItem.getMember().getEmail();
+            if (!memberEmail.equals(cartItemMemberEmail)) {
+                throw new CartItemNotMatchWithMemberException("카트 상품의 소유자와 맞지 않습니다.");
+            }
             cartRepository.delete(cartItem);
         }
-        Long memberId = cartItemsMemberDto.getMemberId();
-        Optional<Member> optionalMember = memberRepository.findById(memberId);
+
+        Optional<Member> optionalMember = memberRepository.findByEmail(memberEmail);
         Member member = optionalMember.orElse(null);
         if (member == null) {
             throw new NoMemberException("해당하는 멤버가 없습니다.");
         }
 
-        return cartRepository.findCartItemsByMemberId(memberId);
+        return cartRepository.findCartItemsByMemberId(member.getId());
     }
 
-    @Override
-    public Page<CartItemDto> getAllCartItemsNoFetch(Long memberId, Pageable pageable) {
-        Optional<Member> optionalMember = memberRepository.findById(memberId);
-        Member member = optionalMember.orElse(null);
-        if (member == null) {
-            throw new NoMemberException("해당 회원이 없습니다.");
-        }
-        Page<CartItem> page = cartRepository.findByMember(member, pageable);
-
-        return page.map(CartItemDto::new);
-    }
 }
